requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

SQLAlchemy==2.0.30
pydantic==2.7.1

python-multipart==0.0.9
python-dotenv==1.0.1

passlib[bcrypt]==1.7.4
bcrypt==4.1.3

alembic==1.13.1


redis==5.0.4
websockets==12.0
psycopg2-binary
asyncpg==0.27.0